Bugfix - target.inventory loop was skipping items, Bugfix - player.inventory was changed to target.inventory, Made firesuits immune from acid
target.inventory ipairs for loop was skipping items due to table.remove
resorting the array whenever an item was removed.  This caused items to
be skipped because they would be out of position.  To fix this, looped
from the end of the list to the beginning, thus array avoids being
resorted if an item is removed.

The zombie acid attack was targetting it's own inventory (player.inventory) instead of the humans (target.inventory), this has been fixed.

Firesuits in inventory should not suffer acid damage.  This seems like
common sense.