Refactored armor for humans
Removed inv_ID tracking for item armor class, instead it now copies the
condition and name of the armor in the armor_INST and removes the item
from inventory.  Lot easier to manage this way, but in the superclass
armor it has to be split via human/zombie.  Possibly change zombie armor
to be more similiar to human armor?  Unsure.