Bugfix - selecting the corresponding human/zombie from player/target now works properly
For some parts of the code for the attack func, we needed to know which
mob was the zombie and which was the human.  It can be confusing as
player and target can be either, therefore we narrowed down who was who,
using some basic mob checks.

This also assumes that humans cannot attack humans, and vice-versa with
zombies!  Otherwise the code will break.