Completely changed how broadcastEvent works.  Code is more simplier and shorter.
The aim in doing this is to get the project ready to remove getMessage()
descriptions and just insert descriptions of the events in outcomes or
item/skill activiations.  Plus this overhaul should allow msgs to be
sent even when an action hasn't taken place by the player.  (ie.  we
need to send a serverwide msg)