Removed commented out return statements
The prior reason these statements were being returned in older versions
of code was to broadcast the msg to player.log via the old broadcast
code.

* Note - GPS.activate still returns a boolean since it is used for
something in the outcome.movement parrt of the code  (although this will
be removed eventually as well)  Once this is accomplished, Outcome can
be changed to be a non-global module.