Refactored replace for optimization and clean code
Moved anonymous function to module level scope.  Moved replacement_tbl
to module level scope that gets overwritten by the replace function when
it is called.  This enables the indexToStringSort function from only
having to be created once, instead of the previous way the code was
written which created the anonymous function every time the replace
function was called!