Split player into zombie and human classes
THIS IS A BIG BIG CHANGE!  Lots of stuff is going to get broken!  But if
this gets done, the benefits will be huge down the road!

Most of the abilities/items/equipment/stats that are checked right now
ALWAYS have to account if the player is a zombie or a human.  By
splitting zombie and human into their own classes, there shouldn't be a
need to check mob_type as much, as all methods will only be available if
apart of that class!  Ie.  using an item, you won't need to check if the
you are human, since only humans will have methods to use items!

Hopefully that makes sense.  In the meantime, cleaning up this code is
going to be one big clusterfuck!  Aggghh!