Fix boundary panning issue
Apparently a Tiled Rectangle object messes up our total room_width and
room_height in the code.  For the Berry library it does not normally
handle rectangles, and I had a black rectangle setup for shading. So
what this means is I will have to use a sprite that is all black for
shading purposes instead of the rectangle object.