Fix zoom to get tile pos and set as map x/y
Uses localToContent method to translate the coordinates with scaling
enabled.  There is a small problem with a few pixels needing to be
offset that changes by scale size.  Couldn't figure out why it was
happening, but I have it working close enough for me to not have to
worry too much.