Add item_effect scene object that can be created dynamically
This scene object is vital because it allows us to create objects after
they have been loaded in by the map.  Previously such a feat was not
possible.  This also uses TexturePacker sprite sheets to load and
organize images.  The sprite sheets won't appear in the berry tilesets.
Might have to be careful and make sure they are destroyed properly.

It's still a WIP.