Change graphics to use placeholder sprites
Due to the problems dealing with SS13 sprites not all being CC-BY-SA,
I'm going to have to ditch most of the art assets.  Thus, for the time
being I'm going to replace all the sprites with placeholder (ugly) pixel
boxes.  Planning on going through opengameart to look for sprites that I
can use for floors and mobs.