Fix equipment to use isVisible method instead of alpha
I was making the mistake to use alpha (aka transparency) to hide or show an equipment object.  This had some really weird behavior especially if the layer was turned on or off.  The correct variable to set should have been object.isVisible and that's what I will be using from now on.