Fix mob movement not being reset properly
Whenever moveMobsToSpawn was called previously, the old positions would be saved.  This was a problem for hidden players because the player searching would save their last position at the search area.  Since a touch event sends a player back to their last location, this meant a human could be sent into a defender area even if they were an attacker.

This bug was a PIA to track down and figure out.  Glad it's finally over with!!!