Change all outside items (except barricade) to be ruined
Most items will spawn in the ruined state when found outside.  Barricades are the only exception.  Barricades will spawn in a worn state.  If the player is searching a junkyard or carpool the barricade will spawn in a pristine state.

Also changed the setting variable to be stage.  This should make things more specific for understanding the code.